Mostly Complete except API Testing

+++ src/main/java/com/twilight/repositories/OrderRepository.java
@@ -11,7 +11,7 @@ import org.springframework.data.domain.Pageable;
 import java.util.List;
 
 
-public interface OrderRepository extends JpaRepository<@NonNull Order,@NonNull Long> {
+public interface OrderRepository extends JpaRepository<@NonNull Order,@NonNull Integer> {
     List<Order> findByCustomerMobNo(String mobNo, Pageable pageable);
 
 }
